refactor: stats->records
